refactor: move task components to parent folder to maintain consistency
