+++ src/components/task-board/RemovePropertyButton.tsx
